(FACT-1968) add cmake config for UDEV
I'm not sure if this is a bug or not. UDEV is a optional dependency,
like libwhereami. cmake searches for suitable libs if it has a proper
configuration file. This file is missing for UDEV so it will never work.

I tested this successfully. Without this change (fact disks.sda):

{
  model => "TOSHIBA DT01ACA3",
  size => "2.73 TiB",
  size_bytes => 3000592982016,
  vendor => "ATA"
}

With this change:

{
  model => "TOSHIBA DT01ACA3",
  serial_number => "532SD5***",
  size => "2.73 TiB",
  size_bytes => 3000592982016,
  vendor => "ATA"
}