(FACT-1824) Offline CPUs count as physical CPUs
If a CPU is set to offline, the /sys/devices/system/cpu/cpu/topology
folder will be removed, and the topology folder contains informations
needed for processor fact.

Now facter will check if the file that contains informations is readable
before reading it. As the file does not exist for offline CPUs, the file
will not be readable and will not cause the bug.