(maint) Install bundle deps before running specs
Facter has some spec tests which it decides to run based on whether it
has ruby and bundler installed.

If bundler is installed, Facter goes ahead and runs `bundle exec rspec`
which fails if dependencies were not installed beforehand.

Stick a `bundle install` before running the tests to ensure we have the
necessary dependencies installed.