(FACT-2346) Check for `ai_canonname` before comparing
On OS X (and probably more BSD-based systems), calling `getaddrinfo`
with a numeric value does not fill the `ai_canonname` field of the
`addrinfo` struct, causing any facter call that queries the networking
resolver to segfault. This happens if the user has a fully numeric
hostname.

From the `getaddrinfo` manpage:

> If the AI_CANONNAME bit is set, a successful call to getaddrinfo()
> will return a NUL-terminated string containing the canonical name of
> the specified hostname in the ai_canonname element of the first
> addrinfo structure returned.

Side-note: `getaddrinfo` will treat any numeric value as an IP, so
numeric hostnames cannot be reliably looked up. The GNU libc
implementation of `getaddrinfo` always fills the `ai_canonname` field,
so this hasn't been a problem on Linux systems.