(FACT-2489) external facts files log update
When there is no appropriate external facts resolver for a file
in `puppet/cache/facts.d` or other external facts dir, a warning
message is logged. This leads to many warning messages in case
multi OS external facts files are in the same directory.

With this change, a debug log message is  logged instead a warning
in case there is no appropriate external facts resolver selected.