(FACT-2666) Add load_external ruby method
The Facter.load_external(<boolean>) will set a flag in facter that will
cause the run to skip loading of external facts.

This method is usefull when we want to get only some core facts, and we
do not want external facts to be loaded.
Puppet uses this setting when evaluating default settings that require
Facter to provide the core fact value. Loading the external facts is not
needed.