(maint) Rework the OS families hierarchy (#2019)
Solaris > BSD > FreeBSD OS hierarchy was convenient to some point but is
inaccurate and starts to be a pain (e.g. networking facts depends on
types only known to Solaris, breaking FFI on FreeBSD).

Move BSD at the same level of the Solaris OS in the hierarchy.  This
makes quite a lot of facts disapear, but it allows us to re-introduce
them without side-effects from the Solaris facts and remove cruft added
to hide facts that make no sense on FreeBSD.