(maint) Improve FreeBSD version detection
freebsd-version -k returns the installed version of the installed
kernel, but jails might have no kernel installed because they are
running from the host's kernel.  In such a situation, freebsd-version
fails — which is expected — but the other information we are interested
in not gathered.

Rework the resolver so that freebsd-version is run twice, so that we can
still gather the information of the running kernel and installed
userland when no installed kernel is available.