(maint) fix execution tests on windows
no_custom_facts_and_load_path.rb is executing an
`agent.rm_rf(path)` where path contained spaces.
This is not working on windows because `path` needs
to be wrapped between double quotes.

This was working on GitHub Actions because the localhost
connection is using `del` in `rm_rf` method