(FACT-2927) Fix computation of metrics prefix
For speed values whose log10 is not in (3, 6, 9, 12), the old code lead
to an error when trying to stringify prefix[power] (which was nil).

The error surfaced when a linux system reported a cpu speed of less
than 1 GHz:

  puts Math.log10(999_999_999).floor
  8

Not really relevant for processor speed, but the hash 'prefix' now
also handles the case of no metrics prefix at all (just 'Hz').