(FACT-3083) - Add wwn and sn to disks fact resolvers/disk.rb
This PR adds the wwn and sn to the disks fact.
The work around the problem, that different os have different locations
for the files that contain the required data, I fell back to the lsblk
tool.

amend 20211023111300 GMT+2: Updated the commit message and removed the
full path to command to use PATH
amend 20211023112500 GMT+2: Changed wwid to the correct wwn variable
name