(FACT-2950) return root fact if defined
Because the toplevel fact does not exists, root of
structured core facts cannot be overriden by a
custom fact. This PR updates the QueryParser logic
to return the root fact, if that is found in the
loaded facts list.

eg: the list contains both `os.name` and `os` facts,
the result will be the `os` fact.

This new fuctionaly allows completly redefinition
of core facts. Eg:

```
❯ cat custom_facts/my_custom_fact.rb

Facter.add(:os) do
  has_weight(999)
  setcode do
    {"architecture"=>"x64"}
  end
end

❯ bx facter os
{
  architecture => "x64"
}

❯ bx facter os.name

❯ bx facter os.architecture
x64
```