(FACT-1847) Fix cloud fact implementation for Azure
Since we now have an az_metadata fact, we can reimplement the cloud fact
using this. The previous implementation was Linux-only and worked by
searching for a specific string in the DHCP lease file, which does not
work anymore (at least not on Ubuntu 20.04 Azure VMs).

The new implementation should work regardless of OS, as it just checks
for the existence of the az_metadata fact.