(FACT-2952) Align the lsbmajdistrelease fact with Facter 3
The `lsbmajdistrelease` fact in Facter 4 was not showing the correct
value on Ubuntu due to their unique approach on versioning.
Example:
In version `20.04.2`, they consider as follows:
 - `20.04` = major version
 - `2`     = minor version

Most of the other Linux operating systems respect the generic semantic
versioning rules (`major.minor.patch`). This commit aligns the fact's
output with Facter 3 accordingly.