(FACT-2998) Fix partial fact matching
Previously, when querying incorrect facts that only match partially
(such as `os.name.test` or `mountpoints./.asd`) would either show
an incorrect value or raise an error.

Match the behavior of Facter 3, which is not to output anything in this
case.

This commit fixes two issues:
1. Querying something like `os.name.test` would return the value of
   `os.name`. This is because resolved facts were incorrectly matched
   with the searched fact and is fixed in the `FactAugmenter`.
2. Querying something that does not exist in the fact collection (such
   as `mountpoints./.asd`, where the `asd` part is missing) would raise
   an error when digging in the collection. This was fixed in the
   `FormatterHelper`, and the digging mechanism was unified in the
   `FactCollection` class.

This commit also changes how `Facter.value` retrieves a fact, matching
the behavior from `FormatterHelper`.

This issue affected access to both core and external facts via CLI and
`Facter.value`.