(FACT-3023) xen command: avoid xen-toolstack warning
xen-toolstack is deprecated and so on Puppet runs in Debian buster you
receive this nettlesome message:

  warning: something called deprecated script /usr/lib/xen-common/bin/xen-toolstack

Prior to this change we used xen-toolstack if it is present, but that is
only necessary if more than one stack is installed. Instead check if we
have multiple tool stacks and if we do use xen-toolstack if it is
present.