(FACT-3059) Fix AIX reporting `odd number of arguments for Hash`
Previously, the aix `info_extractor` would make the assumption that keys are
separated from values with multiple spaces, but in some instances there would be
only one space.

This commit fixes this issue by searching for each known key
(from the PROPERTIES variable), and finds the value, then ignores the rest of
the line. The algorithm repeats for each key.