(FACT-3057) Downcase environment facts
Previously, environment facts were not downcased prior to being added to
the fact collection. This contradicts the Facter 3 behavior[1], and
breaks some use cases where the fact name is upcased when added to the
collection and downcased when resolved via the CLI/Ruby API.

Downcase environment facts before adding them to the collection, and
downcase the user queries passed to `Facter.fact` and `Facter.value`,
which matches the behavior of Facter 3.[2]

Link to the Facter documentation in the README, where this behavior will
also be documented.

[1] https://github.com/puppetlabs/facter/blob/f123fc0c371114e4c6bc61ab245cdb3a32cf5ebb/lib/src/facts/collection.cc#L257
[2] https://github.com/puppetlabs/facter/blob/f123fc0c371114e4c6bc61ab245cdb3a32cf5ebb/lib/src/ruby/module.cc#L483