(FACT-3073) Fix timeout option processing for Facter::Core::Execution.execute
The documentation of Facter::Core::Execution.execute indicate that the
options hash accept 3 values (`expand` is not documented):
  - `on_fail`
  - `logger`
  - `time_limit`

However, when options are extracted by the extract_options method, the
key `limit` is checked instead of `time_limit` as documented.

Finally, previous versions of facter used a `timeout` parameter.

Fix these inconsistencies by using a single name `timeout` consistently.

Adjust the logic so that the documented name that was previously not
working (:time_limit) and the actual name which was working (:time) also
work the same way.