(FACT-3067) Fix resolution of custom facts that partial match legacy fact names
Before this commit, custom facts that had names which partially matched
core legacy facts were being resolved as expected in CLI but not with
the API. Due to the nature of how legacy facts are defined in Facter 4,
with the regular expression used in `FACT_NAME`, `Facter::FactManager` was
not resolving custom facts if core or external facts had any results.

This fix adds a new method to check if the user query matches the name
of already resolved facts to decide whether or not we should also
resolve custom facts.