(FACT-3079) Caching custom facts not working propperly
When using `Facter.value(:fact_name)` inside a custom fact, caching will not
work propperly, as it will caches the value from Facter.value(:fact_name)
instead of the actual fact.

Now it will correctly cache the required fact, by skipping writing the cache
file when it is not required.