(FACT-3094) Aix mountpoint resolver couldn't resolve nfs
This commit fixes NFS mounts being resolved incorrectly by facter. This
happened because the node from where the remote directory is mounted
would be under the `node` header of the mount output. To fix this, we
shift the array by removing the first element, so that every other
column falls into place