(maint) PA-4234 Update Networking Facts test
In GCE the VMs we use can have a `networking.netmask` of 255.255.255.255.
This causes failure in the networking_facts.rb, and networking_facts_with_secondary_ips.rb test, this PR updates the regex
to allow for such an IP.
Also we need to have some small changes to the dmi test because if we use a Google Cloud VM
some default DMI settings are changed.