(maint) Modifies pending test conditional
An issue with earlier versions of timecop caused the 'when yaml
contains Date formatted fields' test to fail with the version of
Psych included with Ruby 3.2.0.

With the release of timecop 0.9.6, this issue has been resolved
and we can remove the test from its conditial pending state when
that version of timecop is present.