(FACT-3170) Add logic to Facter.value and Facter.fact to use cache
Looks like there was something overlooked when it comes to caching with the
value and fact methods. This small fix will make it so we use the cache instead
of looking up the value again.

Also fix up a test that was using reset, instead of clear, in order to clear the
cache. Reset doesn't reset the `already_searched` array.