(maint) Update expected arch for macOS ARM
Ruby 3 reports the architecture as `arm64-darwin` whereas earlier ruby
versions reported `aarch64-darwin`. Since the test runs in both 7.x and
main, we have to handle both. And since the arch must match the directory
in which native extension libraries are installed, explicitly test for
the correct value instead of one regex for all possibilities.