(maint) Don't stub integration tests
Facter's spec_integration tests sometimes fail, because the ffi mocks don't
implement all of the methods for the underlying ffi classes. For example,
FFI::Struct#to_ptr is missing, yet that method is called by
Facter::Util::Resolvers::Ffi::Hostname.getffihostname.

This commit moves the integration tests to a separate GH action. The test
doesn't stub anything, unlike the unit tests. The integration test matrix is
restricted to ruby versions that we currently use on agent and server.

The `ffi` gem is added as an optional dependency to the `integration` group.
This way it remains a soft dependency/feature as not all hosts have compilers.