(FACT-3194) Difference in Facter 3 & 4 ordering
In Facter 4.x, Facter.to_hash() and Facter.resolve() return hashes whose keys
are sorted differently depending on the parameters or which of the two  methods
is called. In Facter 3.x, the hashes are always returned in a stable,
consistent order when using either method.

This commit ensures the returned hashes from Facter.resolve() and
Facter.to_hash() have a stable order by using Facter::Utils.sort_hash_by_key().
Additionally, spec tests are added to test the ordering of the returned hashes.