(maint) Check for valid encoding from spawned process
When facter spawns a new process, it is possible that the encoding
for the output will come back as a non-UTF8 string; for instance,
when the locale is set to CP932 for Japanese. This can occur when
the shell injects a string into the output, such as for a command-
not-found type of error. In these cases, we should fall back to
the external default encoding, otherwise ruby will error on string
manipulations of strings that re not UTF-8.