(FACT-3136) Split instance and project sshKeys on newlines
GCE returns ssh keys as a string with newlines separating each key. When parsing
the project and VM instance data, convert to an array of keys like Facter 3 did.

Note Facter 3 used a streaming JSON parser and it matched all hash keys named
"sshKeys" (at any level), whereas we are only matching the keys for the project
and VM instance, but not custom fields that happen to be called "sshKeys".