(FACT-3186) Handle partitions fact when none exist
Prior to this commit, when a diskless machine used Facter the
partitions fact errored out with an undefined method error. This was
because the machine did not have any partitions and Facter was unable
to dig into the partitions hash because the hash didn't exist.

This commit updates Linux's partition fact to return nil early in
the partitions method if the partitions resolver returns nothing.