(FACT-3178) Ensure Facter escapes metacharacters when searching for fact
Previously, when searching for a fact that matches the user query, if the fact
name contains any metacharacters like dots (.) then the regex comparing the
fact_name and processed_equery_fact would incorrectly match and result in the
wrong fact being resolved.

For example, when running `facter ldom.domainrole.impl`, the legacy alias fact,
ldom_domainrole_impl is returned instead of the ldom structured core fact. This
leads to issues later when extracting the fact value because you cannot dig into
ldom_domainrole_impl and nil would be returned as its value.

This commit ensures any special metacharacters are escaped in the regex so the
correct fact is found and resolved and an rspec test is added to test this
behavior.