(FACT-3089) Retain full uname output for ISA
Previous to this commit, the processor instruction set architecture
(ISA) fact on Linux would split the output for `uname -p` on periods
and return only the last value. Depending on the Linux distribution
and hardware, this could result in unexpected output.

This commit changes the ISA fact on Linux to return the full output
from `uname -p`, which is in line with how this fact is handled on
other *nix operating systems (i.e. macOS, AIX, Solaris).