(FACT-3199) Escape regex special characters for queries
When looking up a query for a fact name, the fact name itself is
plopped into a regular expression to determine if it was a matched
correctly. This caused problems because invalid regex expressions
would cause unexpected stack traces within Facter's evaluation. The
change needed for this was introduced in 6854f6f, so this commit
just adds an additional test to ensure that the invalid regex
error just not occur.