(FACT-3422) Fix SLES SAP use of os distro
In facter 3.x when determining the distro of an OS we looked at the os-release
file's description. The regex we used was good enogugh to capture multiple versions
of SUSE Enterprise Server.
In facter 4.x we switched from reading the description setting to reading the 'id' setting.
This PR adds a bit of logic to that to take into account sles_sap.