(FACT-3423) Re-add development dependency
In 19c2bb1, we moved sys-filesystem from a development to runtime
dependency. This was done because we mistakenly believed that a
LoadError introduced in Facter 4.4.2 was caused because sys-filesystem
was not a runtime dependency.

The LoadError was actually caused by another change: updating the base
resolver to log an error instead of a debug message on LoadError
(50bbef9).

The team that originally wrote Facter 4 intended for sys-filesystem to
be a development dependency due to its dependency on ffi. ffi contains
native extensions and would mean that `gem install facter` would require
a compiler.

This commit moves sys-filesystem back from a runtime to development
dependency and leaves a comment in the gemspec explaining the reasoning
for its placement.