(FACT-3423) Rescue mountpoint resolver LoadError
In 50bbef9, we changed the base resolver's behavior to cause LoadErrors
to log an error by default. Previously, we had only logged a debug
message.

This commit updates several mountpoint resolvers to rescue LoadError
when calling the `read_mountpoint_stats` method, which requires the
sys-filesystem gem. sys-filesystem is only listed as a development
dependency (see dbc5a7c) and may not be present at runtime.