(FACT-3423) Add ffi as development dependency
The ffi gem is a dependency for the FFI module in Facter. However, the
dependency on the ffi gem is only met as a transitive dependency through
sys-filesystem.

This commit adds the ffi gem as a development dependency to the gemspec.

The ffi gem is being added as a development dependency, as opposed to a
runtime dependency, because it contains native extensions. The original
intent was to allow users to install Facter without a compiler to be as
portable as possible, and native extensions would prevent that from
happening.