(FACT-3423) Gracefully handle absent ffi
The ffi gem is required in many parts of Facter, but is not guaranteed
to be present because it is only a development dependency, not a runtime
dependency.

This commit updates modules that require ffi to add exception handling
for when ffi has not been installed and a LoadError is thrown.