(FACT-3427) Improve custom fact logging
Prior to this commit, the define_resolution method was incorrectly
passing two arguments to Facter::Log#log_exception when raising, causing
and ArgumentError.

This commit updates define_resolution's error handling to provide more
information including the fact name and backtrace information (if
available).

Additionally, this commit enables a test that was pending.