(FACT-3433) Log error if external fact is invalid
Prior to this commit, when Facter attempted to load an external fact
that used a non-hash data type, it would cause a fatal exception.

This commit updates Facter to validate whether an external fact is a
non-empty hash and, if it isn't, logs an error. This brings Facter 4
behavior more in line with Facter 3.

This commit also updates the debug message for structured data facts
that are the wrong filetype or empty to provide more useful debug
information.