(FACT-3435) Guard against non-global zones in DMI
Solaris non-global zones have limited access to system information and
aren't able to perform all of the operations of a global zone.

Because of that when the Solaris SPARC DMI resolver ran in a
non-global zone, it wasn't able to access system information via
`prtdiag`, resulting in unexpected text output that caused the resolver
to error.

This commit updates the Solaris SPARC DMI resolver to return if it
receives unexpected text as a result of prtdiag.