(FACT-3452) Make Xen resolver more strict
Prior to this commit, Facter could misidentify non-Xen systems as being
Xen-based.

In some circumstances, a non-Xen system will have /dev/xvd* files
present as symlinks to other devices. This was observed on an Amazon
Linux 2023 machine, which was running on a Nitro (KVM-based) hypervisor
but symlinked /dev/xvda1 to an nvme device.

This commit updates the Xen resolver to check if /dev/xvda1 is a symlink
to avoid misidentifying if a system is Xen-based.