Refactor wildcard matching
It's wasn't super obvious, but if `fact_with_wildcard` was true, then we
returned true or false depending on whether query_fact didn't match the fact
name exactly or not. Refactor the code to make it explicit.

Some more context, some legacy facts are dynamically generated and their names
contain wildcards, such as the `blockdevice_.*_model` legacy fact. If the user
queries for `blockdevice_sda_model` specifically, then we need to return the
legacy fact. In that case `fact_with_wildcard` will be true and the fact and
query names match exactly, so `found_fact?` will return true.

However, if you query for the similarly named structured fact
`blockdevice.sda.model`, then `fact_with_wildcard` will be false, since we want
to return the structured fact, not the legacy one.