Remove Rubocop TODO action
The Rubocop TODO action is an unmaintained action from a former Puppet
employee that seems to have been silently failing for a while now with
"fatal: ambiguous argument 'HEAD~1': unknown revision or path not in
the working tree".

Recently, the action has started loudly failing because of
incompatbility between Ruby, RubyGems, and bundler.

This commit removes the Rubocop TODO action.