Relax os.macosx.version.full requirement
The os_processors_and_kernel test passed on macOS 12 ARM when
"os.macosx.version.full" was "12.4", but failed when it was "12.6.1".

The "full" fact is populated from the result of `sw_vers -productVersion`. If
the patch component is 0, then the patch component is omitted from "full".
Otherwise, it is included.

It just so happens some of our macOS 12 ARM hosts have a mix of patch 0 and
non-zero. However, the issue isn't limited to just that OS and arch. So relax
the requirement for the "patch" component for all macOS hosts.