Fetch the registry value not the type
Facter's product release resolver called `Win32::Registry#each` and incorrectly
assumed the second parameter was the `value`, when actually it was the `type`.
Note the `each` method is an alias for `each_value`, which yields three
parameters: `name`, `type` and `value`[1]`

The issue wasn't noticed because the code also called `Win32::Registry#[]` to
get the value and the tests incorrectly stubbed the Windows registry behavior.

Commit 2c291fea6acb assumed the second parameter was the `value`, as a result
the `os.windows` facts had a value of 1, which corresponds to `REG_SZ`[2]:

    C:\> facter -j os.windows
    "os.windows": {
      "edition_id": 1,
      ..
    }

This isn't the first time we've had problems with overstubbing in facter, see
6e7970e20a, 86048b561a, 58185501e9

Fixes #2683

[1] https://github.com/ruby/ruby/blob/v3_2_3/ext/win32/lib/win32/registry.rb#L579
[2] https://github.com/ruby/ruby/blob/v3_2_3/ext/win32/lib/win32/registry.rb#L114