networking: Workaround 2693 by ignoring resolv.conf entries starting with dot
Currently, systems without a FQDN can be mishandled by facter for certain `/etc/resolv.conf` content.
This was initially noticed when systemd-resolved was installed on a host without domain.

systemd-resolved stub resolver sets 'search .' as a search domain, which results in the
following hostname/domain/fqdn triplet: foo, ., foo...
See: https://github.com/systemd/systemd/blob/v255/src/resolve/resolv.conf#L19

This is wrong on multiple levels: first, facter does not seem to handle '.' (the root level)
well, and there have been PRs to remove the trailing dot in FQDN as far back as 2012 (PR 200),
leaving user with a convenient, but sometimes ambiguous string that is not fully qualified.

More fundamentally, facter sould not be using 'search' or 'domain' in resolv.conf to begin with,
as those are client resolution options, and usually set for convenience, but are not part
of the hostname. However, removing this lookup is much more involved, as facter has been
doing this for years.

Hence, this commit implements a a middle ground solution to support top-level/domainless servers
without making a breaking change.