Ignore proxy for cloud metadata retrieval
Prior to this commit when a user set the http_proxy environment
variable, the Net::HTTP#new in the HTTP resolver would use it when
establishing new connections. While this is the desired behavior in
most cases, this also interfered with fetching cloud metadata for
services like AWS, GCP, and Azure.

Facter had previously dealt with this behavior in earlier versions (2.x)
starting with 23bf7e5, but that was lost during Facter's transition to
C++ then back to Ruby.

This commit exposes Net::HTTP#new's p_addr (proxy address) argument and,
by necessity, the preceding port positional argument, and configures
the AWS, GCE, and Azure resolvers to always pass nil as the proxy
address.