Use Set instead of Array for warn & debug messages
Prior to this commit, Facter's logger used Arrays to store warn and
debug messages. When using the warnonce and debugonce options, Facter
would call #include? on the warn and debug arrays.

This commit updates Facter to use Sets instead of Arrays to store warn
and debug messages, making it more performant to search through the
messages (i.e. call #include?).