(FACT-3465) Allow File.exist? and Dir.entries to receive other arguments
The RSpec `allow` method creates a proxy object to wrap the original object.
This way it can intercept calls to the original object that are "allowed".
However, if the proxy receives a method that isn't allowed, then it will raise
an RSpec exception. Several tests were rescuing the RSpec exception, but still
passing (for the wrong reason).

The rule is if you stub an object and it might be called with other arguments,
then you need to add `and_call_original`. This is especially true for
File.exist? because it's called by rubygems.