(FACT-3468) Create Log instance in the base resolver's init_resolver method
Previously, some resolvers eagerly created a Log, while others relied on the
BaseResolver to lazily create it.

Now eagerly create the Log in the BaseResolver's init_resolver method, which all
derived resolvers already call. This preserves the behavior where the derived class'
name is printed in the log message:

    DEBUG Facter::Resolvers::LsbRelease - Command lsb_release ...

This is because `Log.new(self)` refers to the derived resolver class, e.g.
Facter::Resolvers::LsbRelease, not the BaseResolver.

The base_resolver_spec tests were modified because they were testing that
Facter::Log.new was called a certain way. Instead test the behavior that
`BaseResolver.log` method returns an instance of Facter::Log and calling the
method multiple times returns the same object.