(FACT-3468) Don't mutate resolver global state
Previously, many tests set an instance_spy logger on their resolver's `log` class
variable. But only some of those tests cleared the instance_spy at the end.
Therefore, the spy could leak from one test to another, see where one case was
fixed in 06b1d5b12.

To prevent that from happening again, just let the resolver use its log. Since
we're not using instance spies, we have to set the expectation on the log before
calling the method that causes the log to be written to.

One motivation for installing a spy is to prevent messages from being written to
the terminal. But this accomplishes the same thing but installing a global
Logger in spec_helper that just writes to a StringIO in memory.